// src/enums/EstadoLlamadoEnum.ts
export enum EstadoLlamadoEnum {
  publicacionPendiente = 'Publicación Pendiente',
  abierto = 'Abierto',
  bajarCvs = 'Bajar CVs',
  conformacionTribunal = 'Conformación del tribunal',
  cvsCompartidos = 'CVs Compartidos',
  entrevistas = 'Entrevistas',
  psicotecnicoSolicitado = 'Psicotécnico Solicitado',
  psicotecnicoCompartido = 'Psicotécnico compartido',
  pendienteHacerActa = 'Pendiente hacer acta',
  pendienteHacerFirma = 'Acta pendiente de firma',
  pendieteSubidaCDC = 'Subido CDC',
  pendienteSubidaCDACGA = 'Subido COA/CGA',
  finalizado = 'Finalizado',
  eliminado = 'Eliminado',
}